requirements.txt
accelerate==1.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
Django==5.2.7
django-crispy-forms==2.4
filelock==3.20.0
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
psutil==7.1.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0